Migrate some function to the MaterialView
